// angular-biometric-designer/src/app/shared/models/identity/user.dto.ts
export class UserDto {
    id: string = '';
    id_trans: string = '';
    user_name: string = '';
    user_name_up: string = '';
    password: string = '';
    re_password: string = '';
    full_name: string = '';
    phone: string = '';
    email: string = '';
    branch_code: string = '';
    branch_name: string = '';
    status: string = '';
    last_login: string = '';
    roles: string = '';
    auth_stat: string = '';
    created_by: string = '';
    created_date: string = '';
    last_modified_by: string = '';
    last_modified_date: string = '';
    approved_by: string = '';
    approved_date: string = '';
  }
  